FEAT Rigged Kleene Expressions returning Parse Forests.
The "Rigged Kleene regular expressions in Haskell with Parse Forests"
makes me unhappy; I don't have a good theoretical model for the change
I made.  My expectation is that "one" here isn't just one, but also
carries with it knowledge of what that "one-ness" means; the zero/one
relationship sustains as far as the poset is concerned, but the one
value now carries knowledge the semiring needs to assemble the resulting
parse forest.  This is similar to Matt Might's "smart epsilons," but I
haven't gotten there yet.