FEAT Rigged Glushkov Regular Expressions in BOTH.
Fixes the bug in the Haskell implementation of the Glushkov rigged regular
expressions. It was in Sequence.  It's always in Sequence.

Includes Rigged Gluskhov Regular Expressions in Rust.  The results are
pleasing beyond words.  It works, although it's inefficient as hell,
and I'd like to figure out how to get 'add' and 'mul' to work with
references easily.  But for the moment, this is good stuff, and it's
going to be much improved as time goes by.