FEAT Act II, Scene III is now finished.
Experiment 08 repeats much of Experiment 07, only now with more
advanced and "clever" data structures that allow us to extract information
about the input string from the output string.  The "extended" semirings
with indices doesn't really teach me much about how to do the extension,
since I'd already solved that with Experiment 06, trying to unify what
I knew of Might & Adams' work with Fischer, et. al.

On the other hand, it *does* bring to the surface the kind and amount
of cleverness one needs to be able to make regular expressions as powerful
as Henglein wants them to be.