FEAT Streamlined the zero/one/{set} in LeftLong
LeftLong is now a singular datatype with inherent zero/one characteristics.
This streamlines the definition into a single set type declaration.  While
this isn't necessary, it does make some of the work much easier.

The `rig` function now takes an arbitrary input type, which makes rigging
of complex data streams much easier.

The tests have been expanded to include the three test cases from the
paper for LeftLong.  While it's pleasing to see that it works here,
this doesn't help explain why it doesn't work in the Rust version.