A variety of python experiments to rig rust regex.
The two python examples show the results of using either the delta
operator or the nullability annihilator in-line with sequences.
The result appears to be that the delta operator preserves information
about the state of the parse that is critical to post-process extraction
of semantic information from the syntactical analysis.

Might, Adams, et. al. solve this problem with a combinator that treats the
semiring as a *reduction* and "floats" the stored information toward the
head of the parse tree.  This has the simultaneous effect of preserving
the semantic information, processing that information dynamically
in mid-parse (possibly allowing for future progressive parsing), and
shortening the number of hops between the parse head and the point in
the parser where processing the next character starts.

I think their approach is theoretically sound, in that it doesn't break
the mathematical foundations of semiring parsing, but I'll need more
investigation to be sure.