Delta() operator removed in favor of the more dynamic Red().
The Delta() operator was a stopgap for preserving epsilon
information across concatenations.  The Reduction() operator
is more expensive in terms of processing, but far more flexible,
which is why it's used in Derp-3, and why I've chosen to use it
here.

For the time being, the Rep() operator remains as I do not
yet have a clear understanding of how μ-regexps work in this
context.  What I need now is to read ["Free Monads & Near Semirings"]
(https://people.cs.kuleuven.be/~tom.schrijvers/Research/papers/ppdp2015.pdf)
and see if any of the content there is applicable.  They make
the case that the two may be, in some cases, interchangable,
and that doing so would provide a mathematically sound
implementation of Adams' double tree-rotation.  In ["Fun with
Semirings"](http://stedolan.net/research/semirings.pdf), Dolan claims
that Semirings can be used for data flow analysis, which means that
we may be able to universalize this toolkit even to optimizing the
nullability operator.