FEAT Convert 'Sym' to a predicate handler.
Borrowing a trick from Might (I think), I've abstracted Play's symbol
handler to one that takes a predicate.  This handles the first of OR&T's
requirements that Symbol processing be handled as a *class*.  This handling
does not, unfortunately, include their handling of the symbol class for
the empty string.  That, to me, seems fraught, and I need to understand
why.